% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ra.R
\name{fit_ra}
\alias{fit_ra}
\title{Fit Regression Adjusment}
\usage{
fit_ra(
  X,
  Y,
  Zy,
  Ofam = gaussian(),
  treatment.effect = "ATE",
  weights = rep(1, N)
)
}
\arguments{
\item{X}{numeric vector of (0,1) specifiying the treatment variable.}

\item{Y}{numeric vector sepcifying the outcome variable}

\item{Zy}{design matrix for the outcome model to be fitted to each treatment subgroup.}

\item{Ofam}{family function for the outcome model. See \link[stats]{family} for details of family functions.}

\item{treatment.effect}{the treament effect estimand of interest. 
Can be either \code{"ATE","ATT","ATC"} or \code{"All"} to fit all three.}

\item{weights}{an optional numeric vector of ‘observation weights’ to be used in the fitting process.}
}
\value{
List of fit parameters, which is used to derive an object of class \code{teffects} when called by \link[teffectsR]{teffects}.
}
\description{
Fitting function called by \link[teffectsR]{teffects} when the method is set to \code{'RA'}.
Regression adjusment fits a \link[stats]{glm} model for the outcome to each treatment subgroup.
Potential outcomes are derived using these outcome models.
}
\examples{
#generate some data
N = 50
X = rnorm(N)              #confounder
A = rbinom(N,1,plogis(X)) #treatment variable
Y = X+0.5*A               #continuous outcome
Z = rbinom(N,1,plogis(Y)) #binary outcome
df = data.frame(X=X,A=A,Y=Y,Z=Z)

teffect(A~1,Y~X,data=df,method="RA")
teffect(A~1,Z~X,data=df,outcome.family="binomial",method="RA")
}
