% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFglm.R
\name{IF.glm}
\alias{IF.glm}
\title{Influence Functions from Generalized Linear Models}
\usage{
IF.glm(object, x = object$x)
}
\arguments{
\item{object}{A \code{glm} model object (run with \code{x=TRUE}), or a \code{glm.fit} object}

\item{x}{The model matrix. Not required for a \code{glm} model object (run with \code{x=TRUE}).}
}
\value{
Matrix of influence function values.
}
\description{
\code{IF.glm} returns a matrix of influence functions from either a \code{glm} model object (run with \code{x=TRUE}),
 or a \code{glm.fit} object (in which case the user must parse the design matrix, \code{x}).
 The resulting matrix has one row for each observation, and one column for each GLM model parameter.
}
