% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffects.R
\name{teffect}
\alias{teffect}
\title{Treatment Effects for Causal Inference}
\usage{
teffect(
  exposure.formula,
  outcome.formula,
  outcome.family = "gaussian",
  treatment.effect = "ATE",
  data,
  weights,
  method = "AIPW",
  ...
)
}
\arguments{
\item{exposure.formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class) where the left hand side of the
formula contains the binary exposure variable of interest.}

\item{outcome.formula}{an object of class "\code{\link[stats]{formula}}"
(or one that can be coerced to that class) where the left hand side of the
formula contains the outcome variable of interest.
The right hand side should not contain the exposure.}

\item{outcome.family}{family function for the outcome model, can be can be a
character string naming a family function, a family function or the result of
a call to a family function. (See \link[stats]{family} for details of family
functions.)
For methods \code{"IPW"} and \code{"PSMatching"} no outcome model is
required.}

\item{treatment.effect}{The treatment effect estimand. Can be either
\code{"ATE","ATT","ATC"} or \code{"All"} to fit all three.
The method, \code{"PSMatch"}, can only fit the ATE estimand.}

\item{data}{an optional data frame, list or environment
(or object coercible by \code{\link{as.data.frame}}  to a data frame)
containing the variables in the model. If not found in data, the
variables are taken from environment(formula), typically the environment from
which \code{teffects} is called.}

\item{weights}{an optional vector of ‘observation weights’ to be used in
the fitting process. Should be \code{NULL} or a numeric vector.}

\item{method}{The mediation fitting method to be used. Can be either
\code{"AIPW","IPW","RA"} or \code{"PSMatch"}}

\item{...}{arguments to be passed to the matching function when using the
\code{"PSMatch"} method.
See \link[Matching]{Match} for details.}
}
\value{
An object of class \code{teffect} with effect estimates,
estimated standard errors, Wald based test statistics and estimated
potential outcome means.
}
\description{
\code{teffects} is used to fit treatment effect estimands of a binary
exposure, on an outcome, given a set of variables which adjust for
confounding. Estimands included are: the Average Treatment Effect
(\code{"ATE"}), the Average Treatment Effect in the Treated
(\code{"ATT"}), and the Average Treatment Effect in the Controls
(\code{"ATC"}).
}
\details{
Several methods are implemented to adjust for confounding.
Methods based on regression adjustment
(\code{"\link[=fit_ra]{RA}","\link[=fit_aipw]{AIPW}"}) require
that a glm outcome model is specified (See \code{\link[stats]{glm}}).
Methods base on the propensity score
(\code{"\link[=fit_ipw]{IPW}","\link[=fit_aipw]{AIPW}","\link[=fit_psmatch]{PSMatch}"})
require that a binomial glm model is sepcified for the exposure
(propensity score model).
When models are not required, dummy models must speficied, to parse
the appropriate exposure/ outcome variable.
E.g. when using the method \code{"RA"}, then the exposure model might be
\code{A~1}, where \code{A} is the binary exposure variable of interest.
The right hand side of these dummy models is ignored.
}
\examples{
# Example on Generated data
N <- 50
X <- rnorm(N) # confounder
A <- rbinom(N, 1, plogis(X)) # treatment variable
Y <- X + 0.5 * A # continuous outcome
Z <- rbinom(N, 1, plogis(Y)) # binary outcome
df <- data.frame(X = X, A = A, Y = Y, Z = Z)

# Fit AIPW by default
teffect(A ~ X, Y ~ X, data = df)
teffect(A ~ X, Z ~ X, data = df, outcome.family = "binomial")

# Fit IPW
teffect(A ~ X, Y ~ 1, data = df, method = "IPW")
teffect(A ~ X, Z ~ 1, data = df, method = "IPW")

# Fit RA
teffect(A ~ 1, Y ~ X, data = df, method = "RA")
teffect(A ~ 1, Z ~ X, data = df, outcome.family = "binomial", method = "RA")

# Fit PSMatch
teffect(A ~ X, Y ~ 1, data = df, method = "PSMatch")
teffect(A ~ X, Z ~ 1, data = df, method = "PSMatch")
}
