% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffects_aipw.R
\name{teffects_aipw}
\alias{teffects_aipw}
\title{Treatment Effects - Augmented Inverse Probability Weighting}
\usage{
teffects_aipw(PS.formula, O.formula, O.family, data)
}
\arguments{
\item{PS.formula}{\code{\link[stats]{formula}} object for the Propensity Score logistic regression model.
must have the binary treatment variable as a response variable.}

\item{O.formula}{\code{\link[stats]{formula}} object for the Outcome model to be fitted to the treated and untreated strata separately.}

\item{O.family}{link function for the Outcome model, can be either \code{"gaussian"} or \code{"binomial"}}

\item{data}{data frame to be used for fitting}
}
\value{
An object of class \code{teffects}
}
\description{
Short Description
}
\examples{
#generate some data
N = 1000
X = rnorm(N)              #confounder
A = rbinom(N,1,plogis(X)) #treatment variable
Y = X+0.5*A               #continuous outcome
Z = rbinom(N,1,plogis(Y)) #binary outcome
df = data.frame(X=X,A=A,Y=Y,Z=Z)

teffects_aipw(A~X,Y~X,"gaussian",df)
teffects_aipw(A~X,Z~X,"binomial",df)
}
