% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffects_ipw.R
\name{teffects_ipw}
\alias{teffects_ipw}
\title{Treatment Effects - Inverse Probability Weighting}
\usage{
teffects_ipw(PS.formula, O.variable, data)
}
\arguments{
\item{PS.formula}{\code{\link[stats]{formula}} object for the Propensity Score logistic regression model.
must have the binary treatment variable as a response variable.}

\item{O.variable}{the name of the outcome variable as a string.}

\item{data}{data frame to be used for fitting}
}
\value{
An object of class \code{teffects}
}
\description{
Short Description
}
\examples{
#generate some data
N = 1000
X = rnorm(N)              #confounder
A = rbinom(N,1,plogis(X)) #treatment variable
Y = X+0.5*A               #continuous outcome
Z = rbinom(N,1,plogis(Y)) #binary outcome
df = data.frame(X=X,A=A,Y=Y,Z=Z)

teffects_ipw(A~X,"Y",df)
teffects_ipw(A~X,"Z",df)

}
