% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffects_psmatch.R
\name{teffects_psmatch}
\alias{teffects_psmatch}
\title{Treatment Effects - Propensity Score Matching}
\usage{
teffects_psmatch(PS.formula, O.variable, data, te = "ate", caliper = NULL)
}
\arguments{
\item{PS.formula}{\code{\link[stats]{formula}} object for the Propensity Score logistic regression model.
must have the binary treatment variable as a response variable.}

\item{O.variable}{the name of the outcome variable as a string.}

\item{data}{data frame to be used for fitting}

\item{te}{The treatment effect to be fitted. Can be \code{"ate"} for the average treatment effect or
\code{"atet"} for the average treatment effect in the treated population}

\item{caliper}{Optional scalar denoting the caliper which is passed to the \code{\link[Matching]{Match}} function.
Observations which are outside of the caliper are dropped, with a warning.}
}
\value{
An object of class \code{teffects}
}
\description{
Uses the \code{\link[Matching]{Match}} function to perform matching.
}
\examples{
#generate some data
N = 1000
X = rnorm(N)              #confounder
A = rbinom(N,1,plogis(X)) #treatment variable
Y = X+0.5*A               #continuous outcome
Z = rbinom(N,1,plogis(Y)) #binary outcome
df = data.frame(X=X,A=A,Y=Y,Z=Z)

teffects_psmatch(A~X,"Y",df)
teffects_psmatch(A~X,"Z",df)
}
