% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teffects_ra.R
\name{teffects_ra}
\alias{teffects_ra}
\title{Treatment Effects - Regression Adjusment}
\usage{
teffects_ra(O.formula, T.variable, data, te = "ate", O.family = "gaussian")
}
\arguments{
\item{O.formula}{\code{\link[stats]{formula}} object for the Outcome model to be fitted to the treated and untreated strata separately.}

\item{T.variable}{the name of the treatment variable as a string.}

\item{data}{data frame to be used for fitting}

\item{te}{The treatment effect to be fitted. Can be \code{"ate"} for the average treatment effect or
\code{"atet"} for the average treatment effect in the treated population or
\code{"atc"} for the average treatment effect in the control population.}

\item{O.family}{link function for the Outcome model, can be either \code{"gaussian"} or \code{"binomial"}}
}
\value{
An object of class \code{teffects}
}
\description{
Short Description
}
\examples{
#generate some data
N = 1000
X = rnorm(N)              #confounder
A = rbinom(N,1,plogis(X)) #treatment variable
Y = X+0.5*A               #continuous outcome
Z = rbinom(N,1,plogis(Y)) #binary outcome
df = data.frame(X=X,A=A,Y=Y,Z=Z)

teffects_ra(Y~X,"A",O.family="gaussian",data=df)
teffects_ra(Z~X,"A",O.family="binomial",data=df)
}
